+++ gtfs_kit/routes.py
@@ -724,21 +724,13 @@ def geometrize_routes(
         d["geometry"] = so.linemerge(group.geometry.tolist())
         return pd.Series(d)
 
-    if use_utm:
-        lat, lon = feed.shapes[["shape_pt_lat", "shape_pt_lon"]].values[0]
-        crs = hp.get_utm_crs(lat, lon)
-    else:
-        crs = cs.WGS84
-
     return (
         feed.geometrize_trips(trip_ids, use_utm=use_utm)
         .filter(["route_id", "direction_id", "geometry"])
-        # GeoDataFrame disappears here
         .groupby(groupby_cols)
         .apply(merge_lines)
         .reset_index()
         .merge(feed.routes)
-        .pipe(gp.GeoDataFrame, crs=crs)
     )