+++ tests/test_trips.py
@@ -111,7 +111,7 @@ def test_compute_trip_stats():
     trip_stats_2 = compute_trip_stats(feed.convert_dist("ft"), route_ids=rids)
     d2 = trip_stats_2.distance.iat[0]  # mi
     f = get_convert_dist("km", "mi")
-    assert f(d1) == d2
+    assert np.allclose(f(d1), d2)
 
     # Shapeless feeds should have null entries for distance column
     feed = cairns_shapeless.copy()